add basic example deduplication to apl guide
